Updated blend_predict() to ensure order is preserved.
The cell names are stored as characters, meaning that the order changes from 1,2,3,4 to 1, 10, 100, 1000, 1001, etc. This causes a problem when the blended predictions are returned, because the user may assume that (like the rest of the functions) the order is unchanged. Since the user doesn't have a user-controlled primary key they can verify their predictions against, the result is typically a scrambling of results.

By enforcing that the cell names are numeric, the order is preserved and the lookup functions continue to work properly.  Each object (soils, veg, climate) is produced within-package, so the likelihood of a non-numeric cell name can be controlled. Further type checking may be necessary to ensure this, or you can create a novel primary key within the function.

The key is that the order needs to be preserved when the results are returned to the user.