Added fill_unknown_ages() and supporting files.
This function allows users to provide a dataframe with unknown age vegetation classes (like PineU) and have them filled with the best guess of ages by natural subregion. This follows the method that I developed with Brandon Allen for using the models. Users can select which distribution they want to pull from.